add typing to all pages
