+++ components/pages/signIn/SignInPage.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import Router from 'next/router';
 
 import { withApolloClient } from 'lib/withApolloClient';
-import WithAuth from 'lib/auth/withAuth';
+import withAuth from 'lib/auth/withAuth';
+import { PageContext } from 'types/pageContextInterfaces';
 
 import { HOME } from 'config/routes';
 import { NotifierProvider } from 'contexts/NotifierContext';
@@ -10,16 +11,16 @@ import { NotifierProvider } from 'contexts/NotifierContext';
 import DefaultTemplate from 'components/shared/templates/DefaultTemplate';
 import Notifier from 'components/shared/atoms/Notifier';
 
-import SignUpForm from './components/SignUpForm';
+import SignInForm from './components/SignInForm';
 
 import { PageContentWrapper } from './styled';
 
-const SignUpPage = () => {
+const SignInPage = () => {
   return (
     <NotifierProvider>
-      <DefaultTemplate testId="signup-page">
+      <DefaultTemplate testId="signin-page">
         <PageContentWrapper>
-          <SignUpForm />
+          <SignInForm />
         </PageContentWrapper>
         <Notifier />
       </DefaultTemplate>
@@ -27,10 +28,12 @@ const SignUpPage = () => {
   );
 };
 
-SignUpPage.getInitialProps = ({ res, accessTokenManager }) => {
-  if (accessTokenManager.accessToken) {
+// TODO: duplicated code
+SignInPage.getInitialProps = ({ res, accessTokenManager }: PageContext) => {
+  if (accessTokenManager?.accessToken) {
     if (res) {
-      res.redirect(302, HOME);
+      res.writeHead(302, { Location: HOME });
+      res.end();
     } else {
       Router.push(HOME);
     }
@@ -38,4 +41,4 @@ SignUpPage.getInitialProps = ({ res, accessTokenManager }) => {
   return {};
 };
 
-export default withApolloClient(WithAuth(SignUpPage));
+export default withApolloClient(withAuth(SignInPage));
